HackerRankLeaderboardGUI/requirements_full.txt
pandas>=1.3.0
openpyxl>=3.0.0
requests>=2.25.0
pathlib
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
gspread>=5.0.0
oauth2client>=4.1.0
